--- src/main/java/edu/kis/powp/jobs2d/command/CommandImporter.java
@@ -1,5 +1,0 @@
-package edu.kis.powp.jobs2d.command;
-
-public interface CommandImporter {
-    DriverCommand importCommands(String commands);
-}
