+++ src/main/java/edu/kis/powp/jobs2d/command/importer/ICommandImporter.java
@@ -1,0 +1,7 @@
+package edu.kis.powp.jobs2d.command.importer;
+
+import edu.kis.powp.jobs2d.command.DriverCommand;
+
+public interface ICommandImporter {
+    DriverCommand importCommands(String commands);
+}
