+++ src/test/resources/figures/rectangle.txt
@@ -1,3 +1,4 @@
+Rectangle TXT
 SetPositionTo|100|100
 OperateTo|-100|100
 OperateTo|-100|-100
